Added changeTheme method

+++ src/main/java/com/uep/wap/repository/ChartRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChartRepository extends CrudRepository<Chart, Integer> {
+public interface ChartRepository extends CrudRepository<Chart, Long> {
 }
