+++ src/main/java/com/uep/wap/repository/NewsletterRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NewsletterRepository extends CrudRepository<Newsletter, Integer> {
+public interface NewsletterRepository extends CrudRepository<Newsletter, Long> {
 }
