+++ src/main/java/com/uep/wap/repository/ThemeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ThemeRepository extends CrudRepository<Theme, Integer> {
+public interface ThemeRepository extends CrudRepository<Theme, Long> {
 }
